+++ src/Sahasrara/Utility/Discord.hs
@@ -40,6 +40,7 @@ module Sahasrara.Utility.Discord
   )
 where
 
+import Control.Monad.Cont (liftIO)
 import Control.Monad.Exception (MonadException (throw))
 import Data.Char (isDigit)
 import Data.Foldable (msum)
@@ -54,10 +55,11 @@ import Discord.Internal.Gateway.Cache
 import qualified Discord.Requests as R
 import Discord.Types
 import GHC.Word (Word64)
-import Sahasrara.Internal.Cache
-import Sahasrara.Internal.Embed
-import Sahasrara.Utility (EnvDatabaseDiscord, liftDiscord)
-import Sahasrara.Utility.Exception (BotException (..))
+import System.Environment (lookupEnv)
+import Tablebot.Internal.Cache
+import Tablebot.Internal.Embed
+import Tablebot.Utility (EnvDatabaseDiscord, liftDiscord)
+import Tablebot.Utility.Exception (BotException (..))
 
 -- | @sendMessage@ sends the input message @t@ in the same channel as message
 -- @m@. This returns an @Either RestCallErrorCode Message@ to denote failure or
@@ -231,10 +233,18 @@ getGuildEmoji ename gid = do
 
 -- | search through all known guilds for an emoji with that name
 findEmoji :: Text -> EnvDatabaseDiscord s (Maybe Emoji)
-findEmoji ename = fmap msum (liftDiscord readCache >>= cacheToEmoji)
+findEmoji ename = fmap msum (emojiServers >>= cacheToEmoji)
   where
-    cacheToEmoji :: Cache -> EnvDatabaseDiscord s [Maybe Emoji]
-    cacheToEmoji cache = mapM (getGuildEmoji ename) (keys $ cacheGuilds cache)
+    cacheToEmoji :: [GuildId] -> EnvDatabaseDiscord s [Maybe Emoji]
+    cacheToEmoji ids = mapM (getGuildEmoji ename) ids
+    emojiServers :: EnvDatabaseDiscord s [GuildId]
+    emojiServers = do
+      maybeServers <- liftIO $ lookupEnv "EMOJI_SERVERS"
+      case maybeServers of
+        Just x -> pure (read x)
+        Nothing -> do
+          cache <- liftDiscord readCache
+          pure $ keys $ cacheGuilds cache
 
 -- | Render an Emoji
 formatEmoji :: Emoji -> Text