+++ ChangeLog.md
@@ -1,3 +1,4 @@
 # Changelog for sahasrara
 
 ## Unreleased changes
+- Allow configuration of which servers are allowed to provide emoji